The program itself is very simple.
It's a very simple program - it works by analyzing window titles and muting Spotify if/when it detects that an ad is playing. When an ad plays, the title of the Spotify window is renamed "Spotify", whereas if a song is playing, the window is renamed to the song title/artist name.